Skip after prepare hook
After this change https://github.com/NativeScript/nativescript-cli/pull/2983 the NativeScript CLI can prepare only the JaveScript part of the project.
This is required for cloud builds in the NativeScript Sidekick -> https://www.nativescript.org/nativescript-sidekick.
This means that the native files required in the after prepare hook can be missing.
To check if the native project is prepared we should check the value of nativePrepare.skipNativePrepare which is passed in the hookArgs.